mv role-instance to new dir, test pass
