+++ src/__init__.py
@@ -1,5 +1,6 @@
 __all__ = "User"
 
 from .users.models import User
+from .roles.models import Role
 from .database import Base
 from .database import metadata
